+++ CHANGELOG.md
@@ -15,6 +15,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Add `bigquery.Client` class.
 - Add class attributes `Table.query`, `Table.schema`.
 
+### Changed
+
+- In `Table` and `Topic`, the project ID is no longer changed away from what the user provided.
+  It was more confusing and dangerous than it was helpful.
+
 ## \[v0.3.9\] - 2024-07-02
 
 ### Changed