+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ### Changed
 
 - Reorganize and update data listings.
+- Add FAQ content.
+- Clean up docs. Remove 'TODO's. Add autosummary to module pages.
 
 ## \[v0.3.10\] - 2024-07-22